Deprecate Combine/animation-scheduling effects
