--- Sources/ComposableArchitecture/Effects/TaskResult.swift
@@ -1,1 +1,0 @@
-
